src/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.3
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.1
pandas==1.4.3
Pillow==9.2.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
requests==2.28.1
six==1.16.0
urllib3==1.26.11
Werkzeug==2.2.1
zipp==3.8.1
